Implement basic combat interfaces and hitbox system
- Introduced IFighter interface for managing active hitboxes.
- Introduced IDamageable interface for entities that can receive damage.
- Implemented damage-receiving logic in Skeleton, Sheep, and Player.
- Implemented attacking logic in Player.
- Began draft implementation of hitbox–entity intersection and collision handling.

+++ Barbarian/Header Files/DynamicEntity.h
@@ -11,6 +11,6 @@ protected:
 
 public:
 	DynamicEntity(sf::Image img, std::string name, sf::FloatRect& rect, std::vector<Object*> objectsToInteractWith);
-	virtual void update(double time) override = 0;
+	virtual void update(float time) override = 0;
 	virtual void draw(sf::RenderWindow& w) = 0;
 };
